% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{as.param}
\alias{as.param}
\title{Coerce an object to type 'param'}
\usage{
as.param(x, warn = TRUE)
}
\arguments{
\item{x}{An object}

\item{warn}{TRUE if warning should be thrown if object is not of type param}
}
\value{
An object of type 'param'
}
\description{
Coerce an object to type 'param'
}
\examples{
as.param(param.discrete(c('heads', 'tails')))
as.param(NULL)
as.param(5)
as.param(rnorm, warn=FALSE)

}

