% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_seibert.R
\name{breed.use1}
\alias{breed.mutate}
\alias{breed.rinterpolate}
\alias{breed.rinterpolate.param.discrete}
\alias{breed.rinterpolate.param.distr.int}
\alias{breed.rinterpolate.param.distributed}
\alias{breed.rinterpolate.param.fixed}
\alias{breed.rinterpolate.param.real}
\alias{breed.use1}
\alias{breed.use2}
\title{Parameter breeding functions}
\usage{
breed.use1(param, value1, value2)

breed.use2(param, value1, value2)

breed.mutate(param, value1, value2)

breed.rinterpolate(param, value1, value2)

\method{breed.rinterpolate}{param.real}(param, value1, value2)

\method{breed.rinterpolate}{param.discrete}(param, value1, value2)

\method{breed.rinterpolate}{param.distributed}(param, value1, value2)

\method{breed.rinterpolate}{param.distr.int}(param, value1, value2)

\method{breed.rinterpolate}{param.fixed}(param, value1, value2)
}
\arguments{
\item{param}{An object of type 'param'}

\item{value1}{The first parameter value}

\item{value2}{The second parameter value}
}
\value{
A vector of valid parameter values
}
\description{
Parameter breeding functions
}

