% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{initial.value}
\alias{initial.value}
\title{Generate an initial parameter value}
\usage{
initial.value(param, n = 1, seed = NULL)
}
\arguments{
\item{param}{An object of type 'param'}

\item{n}{The number of initial values to generate}

\item{seed}{The random seed}
}
\value{
A vector of length \code{n} with valid parameter values
}
\description{
Generate an initial parameter value
}
\examples{
p <- param.discrete(c('heads', 'tails', 'edge of coin'),
                    weights=c(0.45, 0.45, 0.1))
initial.value(p, n=10)

}

