% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.param}
\alias{is.param}
\title{Test if an object is a param}
\usage{
is.param(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object is a 'param', FALSE otherwise
}
\description{
Test if an object is a param
}
\examples{
is.param(4) # false
is.param(param.discrete(c('heads', 'tails'))) # true

}

