% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{is.param.valid}
\alias{is.param.valid}
\alias{is.param.valid.default}
\alias{is.param.valid.param.discrete}
\alias{is.param.valid.param.distr.int}
\alias{is.param.valid.param.distributed}
\alias{is.param.valid.param.fixed}
\alias{is.param.valid.param.real}
\title{Validate parameter values}
\usage{
is.param.valid(param, value)

\method{is.param.valid}{default}(param, value)

\method{is.param.valid}{param.discrete}(param, value)

\method{is.param.valid}{param.real}(param, value)

\method{is.param.valid}{param.distributed}(param, value)

\method{is.param.valid}{param.distr.int}(param, value)

\method{is.param.valid}{param.fixed}(param, value)
}
\arguments{
\item{param}{An object of type 'param'}

\item{value}{A vector of values to validate}
}
\value{
A logical vector indicating if \code{value} was valid.
}
\description{
Validate parameter values
}
\examples{
p <- param.discrete(c('heads', 'tails'))
is.param.valid(p, 'heads') # true
is.param.valid(p, 'tails') # true
is.param.valid(p, 'neither heads nor tails') # false

}

