% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param.discrete}
\alias{param.discrete}
\title{Deifine a Discrete Parameter}
\usage{
param.discrete(choices, initial = NULL, weights = NULL, ordered = FALSE,
  ...)
}
\arguments{
\item{choices}{A vector of valid choices. Can be any type.}

\item{initial}{The initial value (must be one of \code{choices} or NULL). See \link{initial.value}.}

\item{weights}{Probability weights for each choice (must be of same length as \code{choices}).}

\item{ordered}{TRUE if order matters, FALSE otherwise.}

\item{...}{Additional arguments to be used in optimisation methods}
}
\value{
An object of type 'param.discrete'
}
\description{
Deifine a Discrete Parameter
}
\examples{
# regular discrete parameter
param.discrete(c("inside", "outside", "I don't care"))

# ordered discrete parameter
param.discrete(c("disagree", "somewhat agree", "agreee"), ordered=TRUE)

# create a parameter that is a loaded coin
param.discrete(c(TRUE, FALSE), weights=c(0.75, 0.25))

}

