% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param.distributed}
\alias{param.distr.int}
\alias{param.distributed}
\title{Define a parameter with a known probability distribution}
\usage{
param.distributed(type = c("norm", "beta", "binom", "cauchy", "chisq", "exp",
  "f", "gamma", "geom", "hyper", "lnorm", "logis", "multinom", "nbinom", "pois",
  "signrank", "t", "unif", "weibull", "wilcox", "birthday", "tukey"),
  args = list(), initial = NULL, ...)

param.distr.int(type = c("norm", "beta", "binom", "cauchy", "chisq", "exp",
  "f", "gamma", "geom", "hyper", "lnorm", "logis", "multinom", "nbinom", "pois",
  "signrank", "t", "unif", "weibull", "wilcox", "birthday", "tukey"),
  args = list(), initial = NULL, ...)
}
\arguments{
\item{type}{The type of probability distrubtion. This is used to call the functions
'rTYPE' and 'dTYPE' (e.g, rnorm and dnorm), which are used by various optimisation
algorithms.}

\item{args}{A named list of arguments to pass to rTYPE and dTYPE (e.g., mean, sd).}

\item{initial}{The initial value or NULL.}

\item{...}{Additional arguments to be used in optimisation methods.}
}
\value{
An object of type 'param.distributed'
}
\description{
Define a parameter with a known probability distribution
}
\examples{
param.distributed("norm", args=list(mean=12, sd=2))
# is the same as
param.normal(mean=12, sd=2)

}
\seealso{
\link{param.normal}, \link{param.real}
}

