% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param.fixed}
\alias{param.fixed}
\title{Define a fixed parameter}
\usage{
param.fixed(value)
}
\arguments{
\item{value}{The value of the parameter, always.}
}
\value{
An object of type 'param.fixed'
}
\description{
Define a fixed parameter
}
\examples{
param.fixed(5)
# can be a complicated value
param.fixed(list(a=1, b=2, c=4))

}

