% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param.normal}
\alias{param.normal}
\alias{param.normal.int}
\title{Define a real parameter with a normal distribution}
\usage{
param.normal(mean, sd, initial = NULL, ...)

param.normal.int(mean, sd, initial = NULL, ...)
}
\arguments{
\item{mean}{The mean of the distribution}

\item{sd}{The standard deviation of the distribution}

\item{initial}{The initial value or NULL}

\item{...}{Additional arguments to be used in optimisation methods.}
}
\value{
An object of type 'param.distributed'
}
\description{
Define a real parameter with a normal distribution
}
\examples{
param.distributed("norm", args=list(mean=12, sd=2))
# is the same as
param.normal(mean=12, sd=2)

}
\seealso{
\link{param.distributed}, \link{param.real}
}

