% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param.real}
\alias{param.real}
\title{Define a Real Parameter}
\usage{
param.real(bounds, initial = NULL, ...)
}
\arguments{
\item{bounds}{A vector of length 2 describing the upper and lower bounds for the parameter.
Some optimisation algorithms can handle values of \code{c(-Inf, Inf)}, but most cannot.}

\item{initial}{The initial value (must be within \code{bounds}), or NULL. See \link{initial.value}}

\item{...}{Additional arguments to be used in optimisation methods}
}
\value{
An object of type 'param.real'
}
\description{
Define a Real Parameter
}
\examples{
# unbounded real parameter (may not work in all optimisation algorithms)
param.real(c(-Inf, Inf))

# bounded real parameter
param.real(c(0, 100))
}

