% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{random.value}
\alias{random.value}
\alias{random.value.param.discrete}
\alias{random.value.param.distr.int}
\alias{random.value.param.distributed}
\alias{random.value.param.real}
\title{Generate a random parameter value}
\usage{
random.value(param, n = 1, seed = NULL)

\method{random.value}{param.discrete}(param, n = 1, seed = NULL)

\method{random.value}{param.real}(param, n = 1, seed = NULL)

\method{random.value}{param.distributed}(param, n = 1, seed = NULL)

\method{random.value}{param.distr.int}(param, n = 1, seed = NULL)
}
\arguments{
\item{param}{An object of type 'param'}

\item{n}{The number of values to generate}

\item{seed}{The random seed}
}
\value{
\code{n} valid random parameter values
}
\description{
Generate a random parameter value
}
\examples{
p <- param.discrete(c('heads', 'tails'))
random.value(p)
random.value(p, n=10)

p <- param.real(c(0, 100))
random.value(p, n=10)

p <- param.normal(50, 10)
random.value(p, n=10)

p <- param.normal.int(50, 10)
random.value(p, n=10)

}

