% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_seibert.R
\name{seibert.breed}
\alias{seibert.breed}
\title{Combine two values of a parameter type using weighted probabilities}
\usage{
seibert.breed(param, value1, value2, prob = list(breed.use1 = 0.41, breed.use2
  = 0.41, breed.rinterpolate = 0.16, breed.random = 0.02), seed = NULL)
}
\arguments{
\item{param}{A 'param' object}

\item{value1}{The first value}

\item{value2}{The second value}

\item{prob}{A named \code{list} in the form \code{functionname=probability}.
The function described by \code{functionname} should take the parameter as
its first argument and (at least) two vector parameters of arbitrary length
and return a vector of identical length. Examples include
\link{breed.use1}, \link{breed.use2}, \link{breed.rinterpolate}, and
\link{breed.random}.}

\item{seed}{A random number seed for replicability.}
}
\value{
A vector of new parameter values
}
\description{
Combine two values of a parameter type using weighted probabilities
}

